feature/HOCContext: Refactored details components, created hoc for person details
