+++ 02.StarDB/03-react-patterns/src/components/hoc-helpers/index.js
@@ -1,3 +1,4 @@
 import withData from './with-data';
+import withSwapiService from './with-swapi-service';
 
-export { withData };
+export { withData, withSwapiService };
