--- 02.StarDB/03-react-patterns/src/components/sw-components/details.js
@@ -1,72 +1,0 @@
-import React from 'react';
-
-import ItemDetails, { Record } from 'components/item-details';
-import { SwapiServiceConsumer } from 'components/swapi-service-context';
-
-const PersonDetails = ({ itemId }) => {
-  return (
-    <SwapiServiceConsumer>
-      {
-        ({ getPerson, getPersonImage }) => {
-          return (
-            <ItemDetails
-              itemId={itemId}
-              getData={getPerson}
-              getImageUrl={getPersonImage}
-            >
-              <Record field="gender" label="Gender" />
-              <Record field="eyeColor" label="Eye Color" />
-            </ItemDetails>
-          );
-        }
-      }
-    </SwapiServiceConsumer>
-  );
-};
-const PlanetDetails = ({ itemId }) => {
-  return (
-    <SwapiServiceConsumer>
-      {
-        ({ getPlanet, getPlanetImage }) => {
-          return (
-            <ItemDetails
-              itemId={itemId}
-              getData={getPlanet}
-              getImageUrl={getPlanetImage}
-            >
-              <Record field="population" label="Population" />
-              <Record field="rotationPeriod" label="Rotation Period" />
-              <Record field="diameter" label="Diameter" />
-            </ItemDetails>
-          );
-        }
-      }
-    </SwapiServiceConsumer>
-  );
-};
-const StarshipDetails = ({ itemId }) => {
-  return (
-    <SwapiServiceConsumer>
-      {
-        ({ getStarship, getStarshipImage }) => {
-          return (
-            <ItemDetails
-              itemId={itemId}
-              getData={getStarship}
-              getImageUrl={getStarshipImage}
-            >
-              <Record field="model" label="Model" />
-              <Record field="length" label="Length" />
-            </ItemDetails>
-          );
-        }
-      }
-    </SwapiServiceConsumer>
-  );
-};
-
-export {
-  PersonDetails,
-  PlanetDetails,
-  StarshipDetails
-};
