+++ 02.StarDB/03-react-patterns/src/components/sw-components/index.js
@@ -1,14 +1,12 @@
+import PersonDetails from './person-details';
+import PlanetDetails from './planet-details';
+import StarshipDetails from './starship-details';
+
 import {
   PersonList,
   PlanetList,
   StarshipList
 } from './item-lists';
-
-import {
-  PersonDetails,
-  PlanetDetails,
-  StarshipDetails
-} from './details';
 
 export {
   PersonList,
