+++ 02.StarDB/03-react-patterns/src/components/sw-components/person-details.js
@@ -1,0 +1,20 @@
+import React from 'react';
+
+import ItemDetails, { Record } from 'components/item-details';
+import { withSwapiService } from 'components/hoc-helpers';
+
+const PersonDetails = ({ itemId, swapiService }) => {
+  const { getPerson, getPersonImage } = swapiService;
+
+  return (
+    <ItemDetails
+      itemId={itemId}
+      getData={getPerson}
+      getImageUrl={getPersonImage}
+    >
+      <Record field="gender" label="Gender" />
+      <Record field="eyeColor" label="Eye Color" />
+    </ItemDetails>
+  );
+};
+export default withSwapiService(PersonDetails);
