+++ 02.StarDB/03-react-patterns/src/components/sw-components/planet-details.js
@@ -1,0 +1,28 @@
+import React from 'react';
+
+import ItemDetails, { Record } from 'components/item-details';
+import { SwapiServiceConsumer } from 'components/swapi-service-context';
+
+const PlanetDetails = ({ itemId }) => {
+  return (
+    <SwapiServiceConsumer>
+      {
+        ({ getPlanet, getPlanetImage }) => {
+          return (
+            <ItemDetails
+              itemId={itemId}
+              getData={getPlanet}
+              getImageUrl={getPlanetImage}
+            >
+              <Record field="population" label="Population" />
+              <Record field="rotationPeriod" label="Rotation Period" />
+              <Record field="diameter" label="Diameter" />
+            </ItemDetails>
+          );
+        }
+      }
+    </SwapiServiceConsumer>
+  );
+};
+
+export default PlanetDetails;
